Changed from float to flex added @media for responsive. Fixed the header //
//- Still trying to learn the proper way to be able to manipulate the flexbox properly.

#about, projects, gallery
@media is going in a column the way I want.
On a normal screen, these sections look good and are doing what I want.
But on a larger monitor, the flex is spreading too wide for my liking. How would I fix it? I tried to put the min-width: 1201px. will need to address later.

Would want coaching on my code, and the way that I have the containers setup, doesn't set well with me.

// Next time around I want to drop the "main-section" to a second page.
I want to try the aesthetics of 1st page for a bit, its typing out the words and the links in the nav bar takes the user to the 2nd page that has the "main-section".

I want to add the typing and deleting sounds to the 1st page.

// I will want to change this competely in the future, as I am 'borrowing' these aesthetics from mcchillin.

// I just learned that styling with #id is not the best practice.